requirements.txt
# fabric==1.8.3
Django==1.6.2
Markdown==2.4
South==0.8.4
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
biopython==1.63
billiard==3.3.0.16
kombu==3.0.13
celery==3.1.10
django-celery==3.1.9
django-extensions==1.3.3
django-nose==1.2
flower==0.6.0
fysom==1.0.15
gunicorn==18.0
hiredis==0.1.2
nose==1.3.1
pandas==0.13.1
# pygraphviz==1.2
python-dateutil==2.2
python-redis-lock==0.1.2
pytz==2014.2
redis==2.9.1
# rpy2==2.3.9
numpy==1.8.1
scipy==0.13.3
scikit-learn==0.14.1
six==1.6.1
tornado==3.2
wsgiref==0.1.2
# MySQL-python==1.2.5
ipdb==0.6.1
ipython==1.1.0
amqp==1.4.4
anyjson==0.3.3
pytest
mongoengine

